separated `TimerTimeline`
